refactor: (WIP) moved common logic from measurement modals to hook
