+++ db.json
@@ -80,6 +80,35 @@
           }
         ]
       }
+    },
+    {
+      "id": "01967724-b982-718f-99ff-1c3080d712ee",
+      "createdAt": 1745755560,
+      "updatedAt": 1745755617,
+      "typeOfMeasurement": "2",
+      "measurement": 44,
+      "afterMealMeasurement": {
+        "meal": [
+          {
+            "id": 0,
+            "portion": 11,
+            "dish": "Борщ",
+            "statistic": {
+              "id": 0,
+              "calories": 44,
+              "proteins": 1.9,
+              "fats": 1.3,
+              "carbohydrates": 6.6,
+              "comment": "The glycemic index is relatively low due to the presence of vegetables and broth"
+            }
+          },
+          {
+            "id": 2,
+            "portion": 33,
+            "dish": "хуй"
+          }
+        ]
+      }
     }
   ],
   "typesOfMeasuremens": [
